Separate SSH part from comm.py
