+++ py-utils/src/utils/comm.py
@@ -15,9 +15,6 @@
 # For any questions about this software or licensing,
 # please email opensource@seagate.com or cortx-questions@seagate.com.
 
-import socket
-import paramiko
-import traceback
 from abc import ABCMeta, abstractmethod
 
 class Channel(metaclass=ABCMeta):
@@ -88,73 +85,3 @@ class Comm(metaclass=ABCMeta):
     @abstractmethod
     def acknowledge(self):
         raise Exception('acknowledge not implemented in Comm class')
-
-
-class SSHSession:
-    """Establish SSH connection on remote host"""
-
-    def __init__(self, host, username, password, port=22):
-        self.host = host
-        self.__user = username
-        self.__pwd = password
-        self.port = port
-        self.timeout = 30
-        self.client = None
-        self.connect()
-
-    def connect(self):
-        """Creates ssh connection
-           Also helps to reconnect same client without invoking new
-           instance again if diconnect caller on the client is called
-        """
-        self.client = paramiko.SSHClient()
-        self.client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-        try:
-            self.client.connect(hostname=self.host,
-                                port=self.port,
-                                username=self.__user,
-                                password=self.__pwd,
-                                timeout=self.timeout)
-        except paramiko.AuthenticationException:
-            raise Exception("Authentication failed, verify your credentials.")
-        except paramiko.SSHException:
-            sshError = traceback.format_exc()
-            raise Exception(f"Could not establish SSH connection. {sshError}")
-        except socket.timeout:
-            raise Exception(f"Timedout. Failed to connect {self.host} in {self.timeout} seconds")
-        except:
-            sshError = traceback.format_exc()
-            raise Exception(f"Exception in connecting to {self.host}\n. {sshError}")
-
-    def exec_command(self, command):
-        """Execute command on remote host
-        """
-        if not self.is_connection_alive:
-            self.connect()
-        stdin, stdout, stderr = self.client.exec_command(command)
-        return (stdin, stdout, stderr)
-
-    def is_connection_alive(self):
-        """Check transporting tunnel is active
-        """
-        if (self.client is None) or (self.client.get_transport() is None):
-            return False
-        return True
-
-    def disconnect(self):
-        """Close the created ssh connection
-        """
-        if self.client:
-            try:
-                # Failure due to this may leave transport tunnel opened.
-                self.client.close()
-            except:
-                pass
-        # Explicitly handling the client connection
-        # instead of relying on garbage collection
-        self.client = None
-
-    def __del__(self):
-        """Destruct the connection
-        """
-        self.disconnect()
